ZM panel revision (v5)
I added 3D models for all the solar bits' footprint, the 8x23 23x25 and 23x42, so that the ZM circuit board could be view in 3D. This allows to check for any errors we couldn't previously visualize along with giving us a sense to what the completed board should look like. Note that the next steps would be Mr. Lyle making sure there are no errors then potentially ordering the board.