Designed 3 possible PCBS for the GPS LNA and antenna circuit board
I made the footprints for the antenna and LNA. I then made a schematic for the circuit board that would contain them. Then I designed 3 separate possible designs for the PCB. One of them has all components on one side, the second spreads all the components along two sides, and the last one only has the antenna on the second side. They are name gps1sideDesign, gps2sideDesign, and gps1.5sideDesign respectively. The smallest by area is the one with just the antenna on the other side (1.5 side). The purpose of making as compact as possible was so there is less disruption with the current solar panel matrices.