obc: rename thermal path for RFM95 to pin 17
There's 4 through holes for soldering the RFM95 module down.
Unfortunately, they are not marked as being on pin17, so kicad
thinks they should be connected to no net.  Now I have marked both
the thermal pad and the thermal holes to be pin 17.

This change only updates the footprint library.